+++ src/test/lambda.cc
@@ -121,8 +121,8 @@ TEST_CASE("reg_lambda")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
-
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/mep.csv").first == MEP_COUNT);
 
   // REGRESSION TEAM OF ONE INDIVIDUAL.
@@ -215,7 +215,8 @@ TEST_CASE("reg_lambda serialization")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/mep.csv").first == MEP_COUNT);
 
   for (unsigned k(0); k < 1000; ++k)
@@ -324,7 +325,8 @@ TEST_CASE("dyn_slot lambda")
 
   constexpr unsigned slots(10);
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/iris.csv").first == IRIS_COUNT);
 
   // DYNSLOT LAMBDA TEAM OF ONE INDIVIDUAL.
@@ -340,7 +342,8 @@ TEST_CASE("dyn_slot serialization")
 
   constexpr unsigned slots(10);
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/iris.csv").first == IRIS_COUNT);
 
   // DYN_SLOT_LAMBDA_F SERIALIZATION - INDIVIDUAL.
@@ -354,7 +357,8 @@ TEST_CASE("gaussian lambda")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/iris.csv").first == IRIS_COUNT);
 
   // GAUSSIAN LAMBDA TEAM OF ONE INDIVIDUAL.
@@ -368,7 +372,8 @@ TEST_CASE("gaussian_lambda serialization")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/iris.csv").first == IRIS_COUNT);
 
   // GAUSSIAN_LAMBDA_F SERIALIZATION - INDIVIDUAL.
@@ -382,7 +387,8 @@ TEST_CASE("binary lambda")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/ionosphere.csv").first == IONOSPHERE_COUNT);
 
   // BINARY LAMBDA TEAM OF ONE INDIVIDUAL.
@@ -396,7 +402,8 @@ TEST_CASE("binary_lambda serialization")
 {
   using namespace vita;
 
-  src_problem pr(initialization::standard);
+  src_problem pr;
+  pr.env.init();
   CHECK(pr.read("./test_resources/ionosphere.csv").first == IONOSPHERE_COUNT);
 
   // BINARY_LAMBDA_F SERIALIZATION - INDIVIDUAL.